Add files via upload
Percent changes of the expected frequency of severe to moderate drought events (SPI ≤ −1.5), calculated at 12- quasi-week (3-month) time scale in Brazil (1980-2024). Red (or blue) shades indicate that the expected frequency of severe to extreme events in 2023 increased (or decreased) in respected to the frequency that would be observed from the stationary approach adopted by the original version of the SPI (6.7%, Table 1).